+++ src/main/java/it/develhope/shoppyz/account/AccountService.java
@@ -7,9 +7,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface AccountService {
-    public Optional<Account> getAccount(Integer id);
+    public Optional<Account> getAccount(Long id);
     public void deleteAccount(Account account);
-    public void deleteById(Integer id);
+    public void deleteById(Long id);
     public void saveAccount(Account account);
     public void updateAccount(Account account);
     public List<Account> getAccountList();
