+++ src/main/java/it/develhope/shoppyz/order/OrderRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-    Optional<Order> findById(Integer id);
+public interface OrderRepository extends JpaRepository<Order, Long> {
+    Optional<Order> findById(Long id);
 
 
 }
