+++ src/main/java/it/develhope/shoppyz/product/ProductRepository.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 
 
 @Repository
-public interface ProductRepository extends JpaRepository <Product, Integer>{
+public interface ProductRepository extends JpaRepository <Product, Long>{
     @Override
-    Optional<Product> findById(Integer id);
+    Optional<Product> findById(Long id);
 
 
 }
